requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.16
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
lxml==4.5.1
numpy==1.18.4
openpyxl==3.0.3
pandas==1.0.3
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
robotframework==3.2.1
robotframework-datadriver==0.3.6
robotframework-excellib==2.0.0
robotframework-seleniumlibrary==4.4.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.46.0
urllib3==1.25.9
webdrivermanager==0.8.0
xlrd==1.2.0
